Move transports to own folder/files for easier dynamic loading. Update tests.
