+++ src/RosLib.ts
@@ -12,13 +12,15 @@ export { default as Topic } from "./core/Topic.js";
 export { default as Param } from "./core/Param.js";
 export { default as Service } from "./core/Service.js";
 export { default as Action } from "./core/Action.js";
+
+// Core Transport exports
 export {
+  AbstractTransport,
   type ITransport,
   type ITransportFactory,
   type TransportEvent,
-  AbstractTransport,
-  WebSocketTransportFactory,
-} from "./core/Transport.js";
+} from "./core/transport/Transport.js";
+export { WebSocketTransportFactory } from "./core/transport/WebSocketTransportFactory.js";
 
 // ActionLib exports
 export { default as ActionClient } from "./actionlib/ActionClient.js";
